Better CLI organization and entrypoint
- Use an entrypoint command instead of a standalone script to
  define the main CLI binary.
- Introduce dev_scripts/tnu to issue CLI commands in a dev env.
- Rearrange CLI code into sensible submodules.
